# dist/openapi.yaml
openapi: 3.0.0
info:
  title: Recipes api
  description: API for recipes project
  version: 1.0.1
servers:
# local:
- url: "http://127.0.0.1:3000"
# remote:
# - url: "http://132.72.65.211"
  description: SwaggerHub API Auto Mocking hey
tags:
- name: Auth
  description: All authentication related routes
- name: Users
  description: All the user's personal app info
- name: Recipes
  description: All reciepes general info
paths:
  /register:
    post:
      tags:
      - Auth
      summary: A new user from the client is added to the server
      description: A JSON object contains all the details about the new user.
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "201":
          description: A new user has been added
        "400":
          description: wrong input parameters
  /login:
    post:
      tags:
      - Auth
      summary: Login an exist user
      description: An exist user login to the system
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginScheme'
        required: true
      responses:
        "200":
          description: |
            User was found in the DB and authenitcated successfully. The session ID is returned in a cookie named `session`.  You should include this cookie in subsequent requests
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: session=87s97sdhgf279293f8w0e9
        "401":
          description: Authentication failed
  /logout:
    post:
      tags:
        - Auth
      summary: logout a user from the system and delete his cookie
      responses:
        "200":
          description: "the user has ben successfully logged out"
  /users/favorites:
    get:
      tags:
      - Users
      summary: Return the favorite recipes that were saved by user
      description: Get an array of recipes that were saved by a specific user
      responses:
        "200":
          description: data about the user's favorites recipes returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePreview'
        "401":
          description: user is not authorized
      security:
      - cookieAuth: []
    post:
      tags:
      - Users
      summary: save recipe in the favorites list of a user
      description: save recipe in the favorites list of a user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reciepeId:
                  type: string
                  description: reciepe's id which the user saved
                  example: 716429
      responses:
        "200":
          description: The Recipe successfully saved as favorite
        "401":
          description: user is not authorized
      security:
      - cookieAuth: []
  /users/seen:
    get:
      tags:
      - Users
      summary: Return the last three recipes seen by the the loggen in user
      description: Get an array of recipes that were seen by a specific user
      responses:
        "200":
          description: data about the user's last 3 seen recipes returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePreview'
        "401":
          description: user is not authorized
      security:
      - cookieAuth: []
    post:
      tags:
      - Users
      summary: save recipe in the seen list of a user
      description: save recipe in the seen list of a user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reciepeId:
                  type: string
                  description: reciepe's id which the user seen
                  example: 716429
      responses:
        "200":
          description: The Recipe successfully saved as seen
        "401":
          description: user is not authorized
      security:
      - cookieAuth: []
  /users/createRecipe:
    post:
      tags:
        - Users
      summary: Add new recipe by the loggen-in user
      description: Add the recipe to the list of the logged-in user list of recipes
      parameters:
        - name: title
          in: body
          description: title of recipes
          required: true
          schema:
            type: string
            example: "Pasta with Garlic, Scallions, Cauliflower & Breadcrumbs"
        - name: image
          in: body
          description: path to image of dish
          required: true
          schema:
            type: string
            example: "https://spoonacular.com/recipeImages/716429-556x370.jpg"
        - name: readyInMinutes
          in: body
          description: time of making the recipes in minutes
          required: true
          schema:
            type: integer
            example: 45
        - name: vegan
          in: body
          description: is the dish vegan 
          required: true
          schema:
            type: boolean
            example: false
        - name: vegetarian
          in: body
          description: is the dish vegetarian
          required: true
          schema:
            type: boolean
            example: false
        - name: glutenFree
          in: body
          description: is the dish glutenFree
          required: true
          schema:
            type: boolean
            example: false
        - name: servings
          in: body
          description: amount of servings made by the recipe
          required: true
          schema:
            type: integer
            example: 2
      responses:
        "200":
          description: Recipe successfully saved to user's list of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePreview'
        "401":
          description: One or more details are missing
      security:
      - cookieAuth: []
  /users/myRecips:
    get:
      tags:
      - Users
      summary: Return list of the user's recipes
      description: Get an array of recipes that were created by a specific user
      responses:
        "200":
          description: user's recipes returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePreview'
        "401":
          description: user is not authorized
      security:
      - cookieAuth: []
  /recipes/{recipeId}:
    get:
      tags:
      - Recipes
      summary: Return preview details on recipe by his ID
      description: Return preview details on recipe by his ID
      parameters:
      - name: recipeId
        in: path
        description: recipe's id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: data about the recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipePreview'
        "404":
          description: recipeId is not exists
  /recipes/random:
    get:
      tags:
      - Recipes
      summary: Return 3 random recipes
      description: Return preview details 3 random recipes
      responses:
        "200":
          description: random recipes returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipePreview'
        "404":
          description: recipes not found
  /recipes/search/:query:
    get:
      tags:
      - Recipes
      summary: Search Recipes
      description: Search recipe using query to filter wanted recipes
      parameters:
        - name: number
          in: path
          description: number of recipes, 5/10/15
          required: true
          schema:
            type: integer
            example: 10
        - name: query
          in: path
          description: Name of the food / recipe wanted
          required: true
          schema:
            type: string
            example: pasta
        - name: cuisine
          in: path
          description: Cuisin(s) of the recipes, seperated by comma
          required: false
          schema:
            type: string
            example: italian
        - name: diet
          in: path
          description: The diet(s) in which the recipes must be suitible, seperated by comma
          required: false
          schema:
            type: string
            example: vegeterian
        - name: intolerance
          in: path
          description: All recipes returned must not contain ingredients enteres, seperated by comma
          required: false
          schema:
            type: string
            example: gluten
      responses:
        "200":
          description: Desired recipes returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipePreview'
        "404":
          description: Search failed
  /recipes/extendInfo/:recipeId:
    get:
      tags:
      - Recipes
      summary: return details of recipe by id
      description: Search recipe using id and get full information of the recipe
      parameters:
        - name: recipe_id
          in: path
          description: id of desired recipe
          required: true
          schema:
            type: integer
            example: 716429
      responses:
        "200":
          description: Desired recipes returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipePreview'
        "404":
          description: Search failed
components:
  schemas:
    User:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: username conatins between 3 to 8 letters only
          example: my_name
        firstname:
          type: string
          description: user's first name
          example: my_first_name
        lastname:
          type: string
          description: user's last name
          example: my_last_name
        country:
          type: string
          description: user's country
          example: Israel
        password:
          maxLength: 10
          minLength: 5
          type: string
          description: password conatins between 5 to 10 chars, at least one of them
            is a number and a special char
          format: password
          example: my_name@12
        email:
          pattern: email
          type: string
          description: user's email
          example: my_name12@gmail.com
        # image-url:
        #   pattern: URL
        #   type: string
        #   description: URL for the user's image
        #   example: https://cloudinary.com/
      description: Model containing user info
    LoginScheme:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: username conatins letters only
        password:
          maxLength: 10
          minLength: 5
          description: password conatins at least one number and one special char
          format: password
      example:
        username: my_name
        password: my_name@12
    RecipePreview:
      type: object
      properties:
        id:
          type: integer
          description: recipe identifier
          example: 102
        title:
          type: string
          description: recipe name
          example: Simple Macaroni and Cheese
        image:
          type: string
          description: recipe image
          format: URL
          example: https://spoonacular.com/recipeimages/655455-566x370.jpg
        readyInMinutes:
          type: integer
          description: the time takes to make this recipe
          example: 45
        popularity:
          type: integer
          description: how many likes got this recipe
          example: 5987
        vegetarian:
          type: boolean
          description: indication if the recipe is vegetarian
          example: false
        vegan:
          type: boolean
          description: indication if the recipe is vegan
          example: false
        glutenFree:
          type: boolean
          description: indication if the recipe is gluten free
          example: false
  